requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
discord==2.1.0
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
joblib==1.2.0
multidict==6.0.4
nltk==3.8.1
openai==0.26.0
pycodestyle==2.10.0
python-dotenv==0.21.0
regex==2022.10.31
requests==2.28.2
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
